+++ DecidedlyShared/Utilities/Geometry.cs
@@ -1,5 +1,7 @@
-﻿using DecidedlyShared.Constants;
+﻿using System.Collections.Generic;
+using DecidedlyShared.Constants;
 using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Graphics;
 
 namespace DecidedlyShared.Utilities;
 
@@ -37,5 +39,27 @@ public class Geometry
     public static Microsoft.Xna.Framework.Rectangle RectToRect(xTile.Dimensions.Rectangle xRect)
     {
         return new Microsoft.Xna.Framework.Rectangle(xRect.X, xRect.Y, xRect.Width, xRect.Height);
+    }
+
+    public static Vector2 GetLargestString(List<string> toMeasure, SpriteFont font)
+    {
+        return GetLargestString(toMeasure.ToArray(), font);
+    }
+
+    public static Vector2 GetLargestString(string[] toMeasure, SpriteFont font)
+    {
+        Vector2 longest = new Vector2();
+
+        foreach (string s in toMeasure)
+        {
+            Vector2 measurement = font.MeasureString(s);
+
+            if (measurement.X > longest.X)
+                longest.X = measurement.X;
+            if (measurement.Y > longest.Y)
+                longest.Y = measurement.Y;
+        }
+
+        return longest;
     }
 }