+++ MappingExtensionsAndExtraProperties/src/Models/FarmAnimals/AnimalModel.cs
@@ -1,0 +1,8 @@
+using System.Collections.Generic;
+
+namespace MappingExtensionsAndExtraProperties.Models.FarmAnimals;
+
+public class AnimalModel
+{
+    public List<Animal> Animals { get; set; }
+}